chore: organize auth platform things to platform/auth/auth.platform.ts
