+++ packages/hoppscotch-selfhost-web/src/platform/auth/auth.platform.ts
