+++ packages/hoppscotch-selfhost-web/src/platform/environments/environments.platform.ts
@@ -1,4 +1,4 @@
-import { authEvents$, def as platformAuth } from "@platform/auth"
+import { authEvents$, def as platformAuth } from "@platform/auth/auth.platform"
 import {
   createEnvironment,
   deleteEnvironment,
