+++ packages/hoppscotch-selfhost-web/src/platform/tabState/tabState.platform.ts
@@ -2,7 +2,7 @@ import { PersistableTabState } from "@hoppscotch/common/services/tab"
 import { HoppRESTDocument } from "@hoppscotch/common/helpers/rest/document"
 import { HoppUser } from "@hoppscotch/common/platform/auth"
 import { TabStatePlatformDef } from "@hoppscotch/common/platform/tab"
-import { def as platformAuth } from "@platform/auth"
+import { def as platformAuth } from "@platform/auth/auth.platform"
 import { getCurrentRestSession, updateUserSession } from "./tabState.api"
 import { SessionType } from "../../api/generated/graphql"
 import * as E from "fp-ts/Either"
